[Feat] #9 BaseView 정의, UIView 객체 상속 및 구현
