+++ MeaningOut/Sources/Presentation/Common/Base/BaseView.swift
@@ -1,0 +1,24 @@
+//
+//  BaseView.swift
+//  MeaningOut
+//
+//  Created by gnksbm on 6/27/24.
+//
+
+import UIKit
+
+class BaseView: UIView {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configureUI()
+        configureLayout()
+    }
+    
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func configureUI() { }
+    func configureLayout() { }
+}
