+++ MeaningOut/Sources/Presentation/Common/Base/BaseViewController.swift
@@ -12,6 +12,7 @@ class BaseViewController: UIViewController {
         super.init(nibName: nil, bundle: nil)
     }
     
+    @available(*, unavailable)
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
