+++ MeaningOut/Sources/Presentation/MainTab/Search/History/View/EmptySearchHistoryView.swift
@@ -9,7 +9,7 @@ import UIKit
 
 import SnapKit
 
-final class EmptySearchHistoryView: UIView {
+final class EmptySearchHistoryView: BaseView {
     private let backgroundImageView = UIImageView().build { builder in
         builder.image(.empty)
             .contentMode(.scaleToFill)
@@ -22,16 +22,7 @@ final class EmptySearchHistoryView: UIView {
             .textAlignment(.center)
     }
     
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configureLayout()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func configureLayout() {
+    override func configureLayout() {
         [backgroundImageView, descriptionLabel].forEach { addSubview($0) }
         
         backgroundImageView.snp.makeConstraints { make in
