+++ MeaningOut/Sources/Presentation/MainTab/Search/History/View/SearchHistoryHeaderView.swift
@@ -9,7 +9,7 @@ import UIKit
 
 import SnapKit
 
-final class SearchHistoryHeaderView: UIView {
+final class SearchHistoryHeaderView: BaseView {
     var removeHandler: () -> Void = { }
     
     private let titleLabel = UILabel().build { builder in
@@ -35,21 +35,11 @@ final class SearchHistoryHeaderView: UIView {
             )
     }
     
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configureUI()
-        configureLayout()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func configureUI() {
+    override func configureUI() {
         backgroundColor = .meaningWhite
     }
     
-    private func configureLayout() {
+    override func configureLayout() {
         [titleLabel, removeButton].forEach { addSubview($0) }
         
         titleLabel.snp.makeConstraints { make in
